[remoteobjects] Use new codepath for Java bridge in instrumentation tests
This CL is parameterizing the Java bridge instrumentation tests to
run with and without using Mojo in JavascriptInjector, running all
the relevant tests and using
JavaBridgeActivityTestRule#MojoTestParams for the ones that pass
and JavaBridgeActivityTestRule#LegacyTestParams for the ones that
fails. So, all tests still run for the old codepath.

Bug: 1105928
Change-Id: Idb3ee3ed4d1cb49fa0fc37d7609ac2e8ef166177
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2348488
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: Oksana Zhuravlova <oksamyt@chromium.org>
Commit-Queue: Miyoung Shin <myid.shin@igalia.com>
Cr-Commit-Position: refs/heads/master@{#798370}