Fix range-percent-intrinsic-size-2 reference
This test worked on Mozilla due to the adoption of orient="vertical"
and the native sizing of the range control. It failed on Chromium
because min-content was less than 30px. This new version appears
to pass on both Chromium and Mozilla.

Fixed: 1108796
Change-Id: Ie5cbaaff6c78f7658931e8d63860383e3501f5ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2355567
Auto-Submit: Mason Freed <masonfreed@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/master@{#798460}