Remove multiplier on launcher search provider groups.
This is the first in a series of changes to how search providers are
created and configured, with the aim of reducing the number of ranking
and filtering steps between the providers and the UI.

Eventually I'd like to repurpose the concept of a 'mixer group'.
This change makes a small start by removing a feature of
groups that we don't use anymore: multipliers on result scores.

This CL also deletes some obsolete tests that don't take into account
modern ranking. This coverage is mostly duplicated by search result
ranker tests.

Bug: 1028447
Change-Id: Ib6d27975ede3b9ee0a9d4b508d262ebff154670b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2342533
Commit-Queue: Tony Yeoman <tby@chromium.org>
Reviewed-by: Jia Meng <jiameng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#798526}