[Autofill][Leipzig] Merging of structured addresses.
This CL adds merging support of structured address tokens.
When two profiles are merged, the substructure that corresponds
to address line that is merged into the new profile is used.
If the feature |kAutofillAddressEnhancementVotes| is disabled,
the structure is dropped in the merging process.

Change-Id: Ifc24c0a65b40fad0ae32b1001cc6560ad946cb6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2353267
Commit-Queue: Matthias Körber <koerber@google.com>
Reviewed-by: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/master@{#798578}