Speculative CHECK to pinpoint crash.
The issue below correlates to element_under_pointer_.at returning a
nullptr. All calls to element_under_pointer_.at but one are prefaced
with an if(element_under_pointer_.Contains(...)) check.
That one call is prefaced with a DCHECK. Most likely culprit is that
there is a real-world case in which the DCHECK doesn't hold and is
not covered by our tests.
Making the DCHECK a CHECK should confirm that. We'd still get the
same number of crashes overall, but with a different stacktrace.

Bug: 1114579
Change-Id: Ie28c971ed2114cf4a780f9eb5366c60d7669315c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2359081
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#798657}