Fix innacurate documentation in TestChromeWebUIControllerFactory
Before the CL, the class-level documentation affirmed that simply adding
a kTestType switch to the CLI would cause this factory to replace the
ChromeWebUIFactory. This was true when it was first introduced [1] but
nowadays it seems that users have to replace the factories by hand (see
the 7 callers of the constructor).

[1] https://codereview.chromium.org/6992019/diff/10003/chrome/browser/ui/webui/chrome_web_ui_factory.cc

Bug: None
Change-Id: Ic9413976b1fd6eb8b1e5ce8c58112629471143c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2360270
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#798711}