Make Chrome conform with other Windows apps in tablet mode.
Chrome now has a visible caption bar in tablet mode above the toolbar/
tabstrip that behaves like most other Windows apps (explorer, console,
etc.)

When maximized in non-split-screen mode:
- Caption bar is thin.
- Restore caption button is present but disabled (greyed out, won't
  highlight). Note that this is an improvement on Windows apps which
  display a non-disabled button that doesn't do anything.
- User can pull down on the caption bar to move the window and go into
  split-screen mode.

When in split-screen mode:
- Caption bar and buttons are default size.
- Maximize caption button is present but disabled (see note above).
- User can pull down on the caption bar to move the window in and out
  of split-screen mode, etc.

This change also makes it much easier for users to generate touch
gestures that clearly show intent to either move the window or open the
tabstrip shade, by dragging on the caption bar vs. the toolbar.

The general accuracy of drag/swipe gestures on the toolbar seems to be
very much improved by this change.

Known issue: when in split-screen Windows sometimes (but not always)
cuts off the top few pixels of a window. This happens to native/first-
party Windows apps the same as Chrome, so is not a Chrome bug, and
because it is intermittent we can't compensate for it except to use a
taller caption bar in this mode (which is also consistent with other
Windows first-party applications).

Note: the ability of the caption button container to be torn out and
moved to the tabstrip container, which was previously required in order
for us to hide the caption buttons and show them when the tabstrip tray
was opened, will be removed in a follow-up CL.

Bug: 1116651
Change-Id: I09ca9381f2add1198a680d9ef6f45cedd7916bf4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2358249
Reviewed-by: Collin Baker <collinbaker@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/master@{#798874}