Fix regression caused by drawing accelerated content on SW canvas
The recent change introducing PaintImage::GetSwSkImage() caused a
regression that occurs when drawing accelerated resources on a software
canvas.

This is a quick fix to revert the culprit caller back to
PaintImage::GetSkImage(). This avoids the cost of multiple readbacks
that GetSwSkImage() incurs. A follow up change will introduce caching in
GetSwSkImage(), at which point we can use it again.

Bug: 1115686
Change-Id: I765f85acf2d53c7f379b4d5cffd86a1224be4318
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2358432
Reviewed-by: Khushal <khushalsagar@chromium.org>
Commit-Queue: Jonah Chin <jochin@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#798898}