Overview: Fix wallpaper dimming when toggling tablet mode
The wallpaper dimming effect isn't applied when the user is in overview
mode and switches from tablet to clamshell mode. This is because normal
and tablet mode have different default wallpaper properties
(blur/opacity). After the properties are applied initially when entering
overview, they are not changed when toggling tablet mode.

This CL fixes this by updating the wallpaper with the correct properties
when starting and ending tablet mode.

Test: manual, added tests
Bug: 1103019
Change-Id: I74fd97f888017f283786bc26ab87ebb39217026c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2340566
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Commit-Queue: Richard Chui <richui@chromium.org>
Cr-Commit-Position: refs/heads/master@{#798899}