Split has_handlers in SetHasTouchEventHandlers to bitfield
In order to allow touch interactions to appropriately work with
composited scrollbars, the scrollbar needs to be designated as a
touch event consumer. This will prevent touch moves on the scrollbar
from being filtered. This requires the browser to know whether a
hit-testable scrollbar is present on a given platform.

This change introduces the mechanism for providing the browser with that
information.

SetHasTouchEventHandlers is converted to SetHasTouchEventConsumers, with
the consumers being JavaScript touch event handlers or scrollbars. The
has_handlers boolean is converted to a bitfield containing these two
pieces of information.

This change does not change the behavior of the browser, and is a
prerequisite to allowing touch dragging and interactions on composited
scrollbars.

Bug: 1082606
Change-Id: I1cf93155ffbccf8dca5196b358f343ee89208abe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2298346
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Bill Budge <bbudge@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Sahir Vellani <sahir.vellani@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#798912}