Add accessibility support for multiselectable aria-activedescendants on Android
This CL adds support for multiselectable aria-activedescendants

For nodes that are multiselectable (e.g. listbox of items with an
aria-activedescendant role), we add more descriptive labels. For
Android R we use the new stateDescription on AccessibilityNodeInfo
object for the given node. To support pre-Android R, we append the
labels to the text for the node.

For a multiselectable node that has no items selected, we add the
text:

"multiselectable, none selected."

For a multiselectable node with n children selected of k total, we
add the text:

"multiselectable, n of k selected."



multiselectable aria activedescendants

AX-Relnotes: Added more descriptive accessibility labels for
Change-Id: I9fe04714c5800939b73b11c3339e628ae7760939
Bug: 1101660
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2341826
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/master@{#798926}