HDR/macOS Tonemapping: Add color conversion
Change the CAMetalLayer that is used to display HDR content to always
be extended-linear-sRGB (and it is already always half-float). Add
code in the shader to support converting extended-sRGB, PQ, and HLG
transfer to linear.

This allows us to promote HLG content to overlays again, because
we are now no longer passing the kCGColorSpaceITUR_2020_HLG color
space directly to the WindowServer (we are interpreting it ourselves).

The next patch in this sequence will add tone mapping support for
HDR-but-only-slightly-HDR displays.

Bug: 1101041, 1079160
Change-Id: Iee02c84d376eacc63fcb4ccd08ef093d2260f284
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2354912
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#798928}