Android: Always offer a zero-suggest verbatim match for websites.
This is done so that the search-ready omnibox (SRO) will be shown
even in cases where there are currently no matches. Specifically,
in Incognito mode tabs.

SRO will be shown whenever the User focuses the Omnibox, or clears
the Omnibox content while visiting a Website, regardless of the
session type.

Bug: 1114125
Change-Id: I4446d10b341d55e9492fc128a4bf61a4edf6d0c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2350762
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#798946}