Invert DCHECK during PDF incremental loading
Fixes a DCHECK in PDFiumEngine::OnPendingRequestComplete() that
incorrectly asserts that the document should be finished loading, when
it actually should assert the opposite.

Updates PDFiumEngineTest::TryLoadIncrementally() to trigger coverage of
this code path. This code path is executed only when pages are in the
pending state when OnPendingRequestComplete() is called, which only
happens when pages become visible outside of a loading operation.

Fixed: 1051548
Change-Id: I035d95c7e626c294a0b3189fcbc6c5ab33944786
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2067743
Reviewed-by: Daniel Hosseinian <dhoss@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: K. Moon <kmoon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#798979}