Make renaming a file not reload the file in chrome://media-app.
Previously we called `launchWithDirectory()` to load in the renamed
file. That blocked the renaming IPC finishing on the entire directory
loading. This change takes a similar approach to SAVE_AS updating
the file in `tokenMap` & `currentFiles` relying on the client app
to reload itself (done in `handleChromiumIPC`) so we don't need to
load the file again!

This change unblocks b/163002455, b/163422302 & b/163432652 as well
as adds a return value for invalid tokens.

It also prevents losing our edit stack when we rename a file!

Bug: b/163422302, b/162281131, b/162281681, 996088
Change-Id: I690963870d1346d31cb77a0c44aeb95802216ef5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2352586
Commit-Queue: David Lei <dlei@google.com>
Reviewed-by: Trent Apted <tapted@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799001}