Caret browsing should persist on startup.
We considered making caret browsing reset on startup, but the
final UX decision was that would just be more confusing.

With this patch, the caret browsing preference persists
like any normal preference - if it was enabled when you
exit Chrome, it will still be enabled the next time you
restart. The only exception is that if you pass the
--enable-caret-browsing switch it will be enabled by
default; this will keep things seamless for users who
have already been using the switch for the past year.

Bug: 1116659
Change-Id: I7653242fa2ee0a929855a3f9f77913a47f9eb9da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2360335
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Auto-Submit: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799030}