Fix CompositeSVG with GraphicsLayer property update
This mainly fixes the crash when we tried to update the property of
a GraphicsLayer that ShouldCreateLayersAfterPaint. The crash could be
fixed by checking if the pending layer's paint artifact is the same as
the passed in paint artifact, but marking the compositing type of the
pending layer can also simplify the code.

Change-Id: I3deef1c60057b1925c7a8d9d891904404f5509bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2360506
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799164}