Rework generated_script handling in MB.
It turns out the work I did on the "wrapped_*" targets in
gn_isolate_map was somewhat overlapping with the already-existing
iOS work to declare things as "generated_scripts". However, the
generated_scripts did not include test_env.py.

So, this CL rearranges things so that the generated script will call
test_env.py, and then we can share logic between the generated_script
targets (and every target on iOS and LaCrOS) and use that as the
basis for the rest of the work going forward.

Bug: 816629
Change-Id: I9d55c3f1587defcb599ddf1df838dd9ed5b6a12c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2357917
Reviewed-by: Erik Staab <estaab@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>
Cr-Commit-Position: refs/heads/master@{#799171}