WebUI Tab Strip: Don't update drag session on dragenter events
Previously, the drag session would be updated as soon as a dragenter
event was fired. This was to allow a dragged out tab to instantly
re-appear in the tab strip once it was dragged back in. Unfortunately,
there exists a bug that causes the drag event's target to be incorrect
if the scale is not 100%.

Removing the update on dragenter should not cause any issues because
the drag session will get updated on the very next dragover event.

Bug: 1117190
Change-Id: Ia06007ab4c3c4827b2783be9cef4df1c30d00f8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2357599
Reviewed-by: dpapad <dpapad@chromium.org>
Commit-Queue: John Lee <johntlee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799219}