GMC: Have the media notification container respond to size changes
This change allows the MediaNotificationContianerImplView to dynamically
respond to size changes of the audio device selector. This will be
helpful in redesign of that UI in which the size of the view will vary.

To facilitate communication between the device picker and the container
a new delegate class was created.

Because the container should support varying sizes, its observer class
was modified to eliminate the notion of expanded vs unexpanded. Instead
observers will be notified of a size change and should query the
container for its new size.

Bug: 1116694
Change-Id: Id832b77a4617645a6c48672670c22ccda88eab82
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2358890
Commit-Queue: Noah Rose Ledesma <noahrose@google.com>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799247}