feed v2: fix call on wrong thread
Added asserts for xsurface APIs that need to be called on the
UI thread.

For processViewAction(), I'm using postTask because the caller does
not call on the UI thread. We can fix this on the other side of
xsurface once it has a postTask API.

Bug: 1117586
Change-Id: I16f298353f7a8f3999215f3aa2e667e7fa282398
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2363352
Reviewed-by: Cathy Li <chili@chromium.org>
Commit-Queue: Dan H <harringtond@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799256}