[omnibox] Pass oft=2 to the search provider for on-clobber ZeroSuggest
This CL enables the passing oft=2 (rather than oft=1) for on-clobber
ZeroSuggest.

oft is a GET parameter for Suggest requests, which means:
  oft=0 normal prefix suggestions
  oft=1 on-focus
  oft=2 on-clobber

We recently created the new oft=2 value, but haven't used it yet.
This is the CL that starts using it.

This is going to temporarily break it, since the server is not yet
set up to handle this yet. We will fix the server next.

Bug: 1106096
Change-Id: I2796dcc9cdb553765876f4341414d15b5e648ad1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2363488
Commit-Queue: Tommy Li <tommycli@chromium.org>
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799267}